customers_service/requirements.txt
flask
mysql-connector-python
pytest
pytest-mock
